refactor: User 도메인 관련 패키지 수정 #6
