+++ src/main/java/com/eduai/auth/jwt/JwtTokenProvider.java
@@ -1,7 +1,7 @@
 
-package com.eduai.auth.security;
+package com.eduai.auth.jwt;
 
-import com.eduai.auth.dto.TokenInfo;
+import com.eduai.auth.application.dto.TokenInfo;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
@@ -10,7 +10,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
