+++ src/main/java/com/eduai/user/presentation/UserController.java
@@ -1,16 +1,20 @@
+package com.eduai.user.presentation;
 
-package com.eduai.auth.presentation;
-
+import com.eduai.common.dto.ApiResult;
+import com.eduai.user.application.dto.UpdateUserRequest;
+import com.eduai.user.presentation.docs.UserApiDocs;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/user")
-public class UserController {
+@RequestMapping("/api/v1/members")
+public class UserController implements UserApiDocs {
 
     @GetMapping("/me")
     public ResponseEntity<String> getCurrentUser(@AuthenticationPrincipal UserDetails userDetails) {
@@ -18,5 +22,11 @@ public class UserController {
             return ResponseEntity.status(401).body("User not authenticated");
         }
         return ResponseEntity.ok("Current user email: " + userDetails.getUsername());
+    }
+
+    @PutMapping
+    @Override
+    public ResponseEntity<ApiResult<Void>> updateMemberInfo(User user, UpdateUserRequest request) {
+        return null;
     }
 }
