+++ src/main/resources/application-local.yml
@@ -18,7 +18,7 @@ spring:
   # --- JPA & Hibernate Configuration ---
   jpa:
     hibernate:
-      ddl-auto: update # Use update to create tables automatically in H2
+      ddl-auto: validate
     show-sql: true
     properties:
       hibernate:
